chore(reports): extracted run-report test block
