+++ libs/bublik/features/run-report/src/lib/run-report-test/index.ts
@@ -1,0 +1,1 @@
+export * from './run-report-test.component';
